Cleaned up and Refactored the Lexer
I have created a lexer before and I know how big one can get especially
if created as a class. The biggest drawback with a class based lexer is
the final size of the Lexer class which I've eliminated with a new
design and also, it gets coupled real fast. Instead, the new design
follows a finite state machine kind of architecture, where the would be
methods of the Lexer class just receive the Lexer state, modify it as
they wish and return control back to the caller, hence reducing coupling
and since they don't have to exist in one source file, the lexer is
modularized which greatly helps in scaling the lexer and adding new
features.
For now, I've cleaned up and I'm only left to write a lexer for numbers
and strings and the lexer phase will be down.