Added proc_loc and refined the Makefile
proc_loc is the location of the token in the processed input string, the
locations differ from the real location since escape characters and
numeric separators are deleted from the input string. Hopefully, the
changes on the Makefile should last longer as the hard coded list of
files to be compiled was getting old real fast.