Merged functions to one construct
To simplify parsing, I made parsing class declarations and function
declarations as simply as calling onto the parse expression which should
yield a Call expression, the called object will be restricted to an
identifier and its arguments the bases for the class and parameters to a
function, but this clashes with the anonymous functions I had started
since parsing anonfn caused an edge case when its body is a paren
enclosed expression which leads to to a Call expression and an error
when collecting the function's body again since the first parse consumed
it. To six this, I have used the fact that brace is not used as an infix
op therefore cannot be recognized by parse expression if it appears
after an expression, other statements from if, to while and with
statements also rely on the open brace not being an expression.

From now on, functions are expressions, but the analyzer will restrict
which syntax is allowed and its rules, etc.