Refactor: Simplify lexer by removing bloat requirements
This projects is blowing up in complexity and some "safe" features are
really not helping. One of them being mutating the input string to
implement features like numeric separators and escapable strings, a
shorter cleaner way around this is to delay the mutation to AST creation
where the resulting nodes like StringNode will receive processed
payloads.