Macro files are getting out of hand, cleanup
This commit has started cleaning up the lexer and called for a
reorganization of the code. The last design is not scaling as I
expected, the token macro files were sweet and short, now big and
cumbersome to work with, and there was, unfortunately, a big hole in the
design, the parser was stuck due to an ambiguity in the grammar making
evolution just hard. This is the first of a series of patches to fix
grammar, mainteinability and extensibility of the project by starting
small.