Tested the new lexer and fixed errors
This was definitely easier compared to the last, the lexer had a few
bugs though, turns out the compiler doesn't eliminate dead code as I
thought, example, consider the expression `a and (B and c)` where a and
c are runtime values while B is a compile time contant or atleast
compile time evaluatable expression, if B evaluates to false, then the
compiler should leave the whole expression to `a` and `a and c` if true,
but this isn't what happens. Check it out later.