SoC: Tokens are a thing by themselves
Due to the great number of tokens available, we cannot have the default
lexer implementation owning them, no, whatever implementation of, the
soon to be added, `ILexer` should lex the tokens provided. However they
do it, tokens are expected with the `Eot` being special and ones
encountered the lexer is assumed to be empty in which all it can produce
is `Eot` tokens.