Refactor the Lexer
Settled on a naming convention which now seems cleaner; `Lexer` instead
of `lexer_t`, `Cursor` instead of `cursor_t`.
The default implementation of `ILexer`, `Lexer`, should only work to
expose a predefined API in which the whole codebase can connect with and
new implementations can conform to seemlessly integrate with the
codebase, this makes testing easier as we can use a `TestLexer` in place
to observe how say the `Parser`, which will implement `IParser` works.