Incompatible changes; Every char should belong to a token
Even though some tokens are utterly useless, they must be generated for
conformance, if we don't, we risk having subtle bugs tucked away along
with the ignored characters. With this, every character in the program
must belong to some token, whatever it is, therefore, the token stream a
lexer produces, when concatenated should yield the original program.
Unfortunately merging with the old codebase is impossible, it was messy,
poorly thought out and pure garbage, the new design eases a bunch of
things which makes the old codebase incompatible with the new one. I've
deleted all old include headers in place of the new.