Refine the API for noexcept
C++ injects a bunch of code to resolve exception handlers; finding
corresponding catch blocks given an exceptions RTTI. This can a bit
heavy especially if no exceptions are expected which are not, atleast
not here. Passing errors is via `IReactor::report(error)` method which
should return a variant of `OnError` which dictates life after error.