Reduce expression node count
Having a node for all possible expressions is awful and laborius to
manage; the expression nodes up to this point had about ten nodes whose
types just inherited from BinaryExpr. Why not merge all to be either
Binary, Unary or have a unique type if necessary, this reduces the node
count greatly. This commit has solved that reducing the expression node
count to twelve approximately making the visitors extremely small in
hindsight, a mistake that killed the last implementation.