Token design is growing uncontrollably
There's more to a token than its lexeme and kind, what about the state
of it? That is, was it passed gracefully for those that posess that
quality like strings, what about the different kinds of strings or
string features like raw, multiline and format strings, if we had kinds
for each then we get 8 different kinds of strings, what if it was not
well formed for some reason like invalid escape sequence, then we need
8 more string kinds to hold the relevant info, discarding this info is
not good since the parser can work with a string expression and accept
it as a literal even though it is not well formed, I mean, better a
broken expression than nothing at all, right? To fix this scaling issue,
we have merged token kinds under umbrellas like String, Number Marker,
Identifier and Operator. All keywords and variable names go under the
Identifier token kind with the specific kind (Variable of keyword) will
be stored in the bit-storage that was used for padding before, now holds
more info about the token, String has features that can be stored as bit
flags like, `is this a raw string?`, `is it well formed?`, `is it a
multiline string?`, the Number token kind umbrellas the different number
types like octal, decimal, float, hexadecimal, binary and error flag,
the Marker houses tokens that are more of markers like the braces,
parens, Error kind, brackets, etc, while the operator captures all
symbolic operators in rattle including but not limited to +, -, *, ,, /,
etc.