Token kind variants implementation
The token variant storage is not random integers stored there, they are
well defined values that can be checked consistenly hence using enums to
defined then, think of the storage as type erased bit-storage for more
info about the token in which the infor is as added in this commit. Some
like the String variants are flags while others like Number are
enumerations for the different kinds of Numbers supported by the
language.