Hide level tracking to prevent invalidation
The syntax analyzer highly relies on level tracking (how deep the
expression is in the expression parser-tree) which can be used to set
constraints on the expression with the root expression having a level of
zero. The level was simply an integer which can accidentally be
modified, so having it hidden behind getter only abstractions makes for
a better tracking system using RAII to ensure proper level exits for
each entry. The statements Continue and Break have been made to events
since they simply say jump to the start or end of a loop which sounds
like an event, there you have it.